eptri: OutFIFOInterface: reset fifo_ready when token is ignored
Currently, fifo_ready is set when a new token arrives to a primed
endpoint number while the FIFO is enabled; and fifo_ready is cleared
once a complete packet has been stored and ACK'd.

This is problematic if there can be cases where the OutFIFOInterface is
ready to receive a packet, and sees a token which is for a primed
endpoint number but which does not actually result in a packet being
stored and ACK'd.

I think there are probably several such types of tokens:

 - SETUP
 - PING
 - IN

When this happens, fifo_ready stays set until the next time a packet
is stored and ACK'd; which is problematic because fifo_ready is supposed
to guard against an endpoint being primed while an incoming packet for
that endpoint is mid-transmission.

To fix this in a way that works correctly for both OUT and PING tokens,
reset fifo_ready when a new token arrives that we're not ready to
handle.